+++ crates/transaction-pool/src/validate/eth.rs
@@ -3,11 +3,8 @@
 use crate::{
     error::InvalidPoolTransactionError,
     traits::{PoolTransaction, TransactionOrigin},
-    validate::{
-        task::ValidationJobSender, TransactionValidatorError, ValidTransaction, ValidationTask,
-        MAX_INIT_CODE_SIZE, TX_MAX_SIZE,
-    },
-    TransactionValidationOutcome, TransactionValidator,
+    validate::{ValidTransaction, ValidationTask, MAX_INIT_CODE_SIZE, TX_MAX_SIZE},
+    TransactionValidationOutcome, TransactionValidationTaskExecutor, TransactionValidator,
 };
 use reth_primitives::{
     constants::ETHEREUM_BLOCK_GAS_LIMIT, ChainSpec, InvalidTransactionError, EIP1559_TX_TYPE_ID,
@@ -16,129 +13,13 @@ use reth_primitives::{
 use reth_provider::{AccountReader, StateProviderFactory};
 use reth_tasks::TaskSpawner;
 use std::{marker::PhantomData, sync::Arc};
-use tokio::sync::{oneshot, Mutex};
-
-/// A [TransactionValidator] implementation that validates ethereum transaction.
-///
-/// This validator is non-blocking, all validation work is done in a separate task.
-#[derive(Debug, Clone)]
-pub struct TransactionValidationTaskExecutor<V> {
-    /// The validator that will validate transactions on a separate task.
-    validator: V,
-    /// The sender half to validation tasks that perform the actual validation.
-    to_validation_task: Arc<Mutex<ValidationJobSender>>,
-}
+use tokio::sync::Mutex;
 
 /// Validator for Ethereum transactions.
 #[derive(Debug)]
 pub struct EthTransactionValidator<Client, T> {
     /// The type that performs the actual validation.
-    inner: Arc<EthTransactionValidatorInner<Client, T>>,
-}
-
-// === impl TransactionValidationTaskExecutor ===
-
-impl TransactionValidationTaskExecutor<()> {
-    /// Convenience method to create a [EthTransactionValidatorBuilder]
-    pub fn eth_builder(chain_spec: Arc<ChainSpec>) -> EthTransactionValidatorBuilder {
-        EthTransactionValidatorBuilder::new(chain_spec)
-    }
-}
-
-impl<Client, Tx> TransactionValidationTaskExecutor<EthTransactionValidator<Client, Tx>> {
-    /// Creates a new instance for the given [ChainSpec]
-    ///
-    /// This will spawn a single validation tasks that performs the actual validation.
-    /// See [TransactionValidationTaskExecutor::eth_with_additional_tasks]
-    pub fn eth<T>(client: Client, chain_spec: Arc<ChainSpec>, tasks: T) -> Self
-    where
-        T: TaskSpawner,
-    {
-        Self::eth_with_additional_tasks(client, chain_spec, tasks, 0)
-    }
-
-    /// Creates a new instance for the given [ChainSpec]
-    ///
-    /// By default this will enable support for:
-    ///   - shanghai
-    ///   - eip1559
-    ///   - eip2930
-    ///
-    /// This will always spawn a validation task that performs the actual validation. It will spawn
-    /// `num_additional_tasks` additional tasks.
-    pub fn eth_with_additional_tasks<T>(
-        client: Client,
-        chain_spec: Arc<ChainSpec>,
-        tasks: T,
-        num_additional_tasks: usize,
-    ) -> Self
-    where
-        T: TaskSpawner,
-    {
-        EthTransactionValidatorBuilder::new(chain_spec)
-            .with_additional_tasks(num_additional_tasks)
-            .build(client, tasks)
-    }
-
-    /// Returns the configured chain id
-    pub fn chain_id(&self) -> u64 {
-        self.validator.inner.chain_id()
-    }
-}
-
-impl<V: TransactionValidator + Clone> TransactionValidationTaskExecutor<V> {
-    /// Creates a new executor instance with the given validator for transaction validation.
-    ///
-    /// Initializes the executor with the provided validator and sets up communication for
-    /// validation tasks.
-    pub fn new(validator: V) -> Self {
-        let (tx, _) = ValidationTask::new();
-        Self { validator, to_validation_task: Arc::new(Mutex::new(tx)) }
-    }
-}
-
-#[async_trait::async_trait]
-impl<Client, Tx> TransactionValidator
-    for TransactionValidationTaskExecutor<EthTransactionValidator<Client, Tx>>
-where
-    Client: StateProviderFactory + Clone + 'static,
-    Tx: PoolTransaction + 'static,
-{
-    type Transaction = Tx;
-
-    async fn validate_transaction(
-        &self,
-        origin: TransactionOrigin,
-        transaction: Self::Transaction,
-    ) -> TransactionValidationOutcome<Self::Transaction> {
-        let hash = *transaction.hash();
-        let (tx, rx) = oneshot::channel();
-        {
-            let to_validation_task = self.to_validation_task.clone();
-            let to_validation_task = to_validation_task.lock().await;
-            let validator = Arc::clone(&self.validator.inner);
-            let res = to_validation_task
-                .send(Box::pin(async move {
-                    let res = validator.validate_transaction(origin, transaction).await;
-                    let _ = tx.send(res);
-                }))
-                .await;
-            if res.is_err() {
-                return TransactionValidationOutcome::Error(
-                    hash,
-                    Box::new(TransactionValidatorError::ValidationServiceUnreachable),
-                )
-            }
-        }
-
-        match rx.await {
-            Ok(res) => res,
-            Err(_) => TransactionValidationOutcome::Error(
-                hash,
-                Box::new(TransactionValidatorError::ValidationServiceUnreachable),
-            ),
-        }
-    }
+    pub inner: Arc<EthTransactionValidatorInner<Client, T>>,
 }
 
 /// A builder for [TransactionValidationTaskExecutor]
@@ -328,7 +209,7 @@ impl EthTransactionValidatorBuilder {
 
 /// A [TransactionValidator] implementation that validates ethereum transaction.
 #[derive(Debug, Clone)]
-struct EthTransactionValidatorInner<Client, T> {
+pub struct EthTransactionValidatorInner<Client, T> {
     /// Spec of the chain
     chain_spec: Arc<ChainSpec>,
     /// This type fetches account info from the db
@@ -357,7 +238,7 @@ struct EthTransactionValidatorInner<Client, T> {
 
 impl<Client, Tx> EthTransactionValidatorInner<Client, Tx> {
     /// Returns the configured chain id
-    fn chain_id(&self) -> u64 {
+    pub fn chain_id(&self) -> u64 {
         self.chain_spec.chain().id()
     }
 }
